+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Helicopter RPM indicator with configurable limits
 - Target IAS indicator
+- Traffic display on HSI
 
 ## [0.8.1] - 2023-04-17